INIT: refactor for contract clients
