+++ src/index.ts
@@ -1,4 +1,6 @@
 export { Account as Account } from './Account';
+export { VaultClient as VaultClient } from './VaultClient';
+export { StakingClient as StakingClient } from './StakingClient';
 export { Vault as Vault } from './Vault';
 export * from './vaultDiscovery';
 export * from './utils';
