+++ src/interfaces/index.ts
@@ -1,0 +1,23 @@
+export interface AbiInterface {
+  sig: string[];
+};
+
+/**
+ * This is passed as an argument to the [[Account]] constructor
+ */
+ export interface AccountInterface {
+  /** @property An optional string of 25 words seperated by spaces that can be used to recover an algorand account */
+  mnemonic?: string;
+  /** @property An optional array of numbers that can be used to recover an algorand account */
+  secretKey?: number[];
+  /** @property An optional name of a signer or wallet provider that can be used to sign transactions, should be set along with the provider property. Only required when not using a mnemonic or secretKey */
+  signer?: string | 'MyAlgoConnect';
+  /** @property The desired network to connect to, the default value is LocalHost */
+  network?: 'LocalHost' | 'MainNet' | 'TestNet';
+  /** @property An optional instance of the provider object for the signer specified */
+  provider?: any;
+  /** @property An optional instance of the reach standard library */
+  reachStdLib?: any;
+  /** @property An optional instance of an account from the reach standard library. Used to reconnect via a frontend */
+  networkAccount?: any;
+}
